Refactored iqdb to scraper, added to wrapper, added configs for everything.
Refactored the name of iqdb to scraper, as it no longer primarily uses IQDB as a source and
scraper is more general as a term.

Added config files and loading methods for wrapper, the wrapper class takes the paths to the
other config files from its config, and then loads them and passes them down to their respective
classes. This will replace commandline args and pipelines as the primary control method.

Added fun to scraper in case more features are deemed needed for massive pervert mode.